fix: 全项目类型检查清零 + 修复若干运行时 bug
tsc --noEmit 此前累计 57 处报错（构建走 esbuild 不跑 tsc 而长期被忽略），
本次全部清零，并修掉其中真正影响运行的若干 bug。全部 369 测试通过、生产构建正常。

真实运行 bug：
- applyAssistantPostProcessing: 小红书回复重试分支漏传 `|| ''`，xsecToken 为空时
  重试会把 undefined 传给 MCP（首调已兜底，重试没兜），改为与首调一致。
- mcdToolBridge / McdMiniApp: 快照 step 缺 'success'，下单完成态无法传给 AI 上下文
  （会显示成 '?'），补全类型与文案。
- vite.config: dev proxy 的 `router` 选项 Vite/http-proxy 根本不支持，海外区域路由
  回调从未生效（被静默忽略）。移除死代码并注明区域路由由生产环境函数处理。
- SongwritingApp: 记忆宫殿检索拼接里读了 SongSheet 上不存在的 `theme` 字段（恒为
  undefined），移除无效读取。

类型健壮性（不改运行行为）：
- MemoryPalaceApp: 17 处 char 访问在“非 globalSettings 视图（1411 行守卫已保证 char
  存在）”下补 `!`，与文件既有写法一致。
- novelUtils: forEach → for...of，修复闭包打断 currentSection 控制流收窄导致的 never。
- ThemeMaker: 捕获 selectorMatch.index 到局部 const；移除与外层重复的 activeTab!=='css' 判断。
- charMusicPersona/groupPipeline/McdMiniApp/SongwritingApp: 修正可空/联合类型与索引签名。
- BankDollhouse: 图标映射改用 Phosphor 的 Icon 类型。

环境/配置：
- tsconfig: 排除 worker/proactive-push（独立打包部署的 Cloudflare Worker，运行时非
  浏览器 DOM，被应用 tsconfig 误检 D1Database/BufferSource）。
- vite-env.d.ts: 为 esm.sh 的 html2canvas URL import 声明模块类型。

+++ utils/memoryPalace/groupPipeline.ts
@@ -266,7 +266,8 @@ export async function processGroupNewMessages(
                     const vector = vectors[i];
 
                     // 与该成员已有记忆去重（同样的群记忆草稿可能跟以前的群记忆撞）
-                    const isDup = existingVectors.some(ev => cosineSimilarity(vector, ev.vector) > DEDUP_THRESHOLD);
+                    // MemoryVectorDB.getAllByCharId 出库即 ensureFloat32，运行时恒为 Float32Array
+                    const isDup = existingVectors.some(ev => cosineSimilarity(vector, ev.vector as Float32Array) > DEDUP_THRESHOLD);
                     if (isDup) {
                         console.log(`♻️ [GroupPalace] ${member.name}：重复群记忆跳过 "${draft.content.slice(0, 30)}..."`);
                         continue;
